Rename logic to lock_calculation
