+++ conder_core/src/main/abstract/mongo_logic/lock_calculation.ts
